installed-tests/testCommandLine: Fix on UTF8-only filesystems
On filesystems which require valid UTF8 filenames, such as ZFS
with the utf8only option turned on, tests involving invalid UTF8
filenames can't work.

Skip the one test which requires such a file in this case.

+++ installed-tests/scripts/testCommandLine.sh
@@ -212,10 +212,13 @@ report "Unicode pathed encoding should work for module run."
 rm -r Код
 
 # non UTF8 file names throws error
-bash -c "touch $'\xff'"
-G_FILENAME_ENCODING=utf8 $gjs -m printFiles.js 2>&1 | grep -q 'Gjs-CRITICAL.*Could not convert filename string to UTF-8 for string: \\377'
-report "Throws error if filename is not UTF8"
-bash -c "rm ''$'\377'"
+if bash -c "touch $'\xff'" 2>/dev/null; then
+    G_FILENAME_ENCODING=utf8 $gjs -m printFiles.js 2>&1 | grep -q 'Gjs-CRITICAL.*Could not convert filename string to UTF-8 for string: \\377'
+    report "Throws error if filename is not UTF8"
+    bash -c "rm ''$'\377'"
+else
+    skip "Throws error if filename is not UTF8" "filesystem doesn't support non UTF8 file names."
+fi
 
 # gjs --help prints GJS help
 $gjs --help >/dev/null
